NodeDefFilePreviewPopper (WIP)
